Collect correct enum base type

+++ tests/Dibix.Sdk.Tests/Resources/CodeGeneration/Enum2.cs
@@ -13,7 +13,7 @@ using Dibix;
 #region Contracts
 namespace Dibix.Sdk.Tests.DomainModel.Policies
 {
-    public enum SecurityPolicy : int
+    public enum SecurityPolicy : short
     {
         Feature1 = 101,
         Feature2 = 102,
